Correct encoding on request to tool API
If encoding is not specified on a request then the default encoding of 'UTF-8"
is passed to `toString` causing the response body to effectively be utf 8 encoded
and mangling any necessary file headers. In cases where a non-file result is
expected, i.e. JSON data, it is probably best to explicitly set the encoding to `utf-8`
for the sake of consistency.

refactored results endpoint to use relative URIs and removed use of base 64 encoded
file data since  it was unneccessary.
Closes #53