adds new debugging options to make ui debugging nicer.  - use `docker-compose -f docker-compose.ui-debug.yaml up -d --build`  - mounts the ui directory as a volume before starting the npm server
be aware that this reinstalls all dependencies at start, which can be
time consuming. If You receive a 502 on `localhost` you may need to
wait longer for dependencies to finish installing and the server to
start. Once the server has started, changes made locally will be
reflected and the server should recompile to reflect changes automatically